+++ src/components/Layouts/Landing/Footer/index.js
