+++ src/components/Layouts/Landing/Footer/template.js
