+++ src/components/Layouts/Landing/style.scss
