+++ src/components/Layouts/Wallet/Header/Ticker/index.js
