+++ src/components/Layouts/Wallet/Header/Ticker/style.scss
