+++ src/components/Layouts/Wallet/Header/WhatsNew/style.scss
