+++ src/components/Layouts/Wallet/MenuTop/ClipboardNextAddress/index.js
