+++ src/components/Layouts/Wallet/MenuTop/ClipboardNextAddress/style.scss
