+++ src/scenes/Login/LoginForm/style.scss
@@ -5,7 +5,7 @@
   max-height: 600px;
   @include box-shadow(0 0 20px -8px black);
   overflow: auto;
-  background-color: white;
+  // background-color: white;
 
   & h3 {
