+++ src/scenes/Login/index.js
@@ -5,6 +5,8 @@ import {bindActionCreators} from 'redux'
 import Login from './template.js'
 import * as authActions from 'data/Auth/actions.js'
 
+import styles from './style.scss'
+
 class LoginContainer extends React.Component {
   constructor () {
     super()
@@ -40,7 +42,7 @@ class LoginContainer extends React.Component {
 
   render () {
     return (
-      <Login onChange={this.onChange} onClick={this.onClick} />
+      <Login className={styles.bgBlack} onChange={this.onChange} onClick={this.onClick} />
     )
   }
 }
