+++ src/scenes/Login/style.scss
@@ -1,4 +1,4 @@
 
 .login {
-
+  background-color: greenyellow;
 }
